requirements.txt
pyserial
Flask==3.1.2

